Cleaned up HomeBody files
